@@ -10,8 +10,10 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -32,22 +34,22 @@
 		"taken"
 	],
 	"dependencies": {
-		"got": "^10.6.0",
+		"got": "^11.8.2",
 		"is-name-taken": "^2.0.0",
-		"is-scoped": "^2.1.0",
-		"is-url-superb": "^4.0.0",
+		"is-scoped": "^3.0.0",
+		"is-url-superb": "^6.0.0",
 		"lodash.zip": "^4.2.0",
 		"org-regex": "^1.0.0",
-		"p-map": "^3.0.0",
-		"registry-auth-token": "^4.0.0",
-		"registry-url": "^5.1.0",
+		"p-map": "^5.0.0",
+		"registry-auth-token": "^4.2.1",
+		"registry-url": "^6.0.0",
 		"validate-npm-package-name": "^3.0.0"
 	},
 	"devDependencies": {
-		"aggregate-error": "^3.0.1",
-		"ava": "^2.1.0",
-		"tsd": "^0.11.0",
-		"unique-string": "^2.0.0",
-		"xo": "^0.26.1"
+		"aggregate-error": "^4.0.0",
+		"ava": "^3.15.0",
+		"tsd": "^0.14.0",
+		"unique-string": "^3.0.0",
+		"xo": "^0.38.2"
 	}
 }
